+++ pkg/models/orders.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Order struct {
-	ID        uint      `json:"id" gorm:"primarykey" faker:"-"`
+	ID        int64      `json:"id" gorm:"primarykey" faker:"-"`
 	CreatedAt time.Time `json:"-"`
 	UpdatedAt time.Time `json:"-"`
-	ClientID  uint      `json:"clientId" gorm:"OnDelete:CASCADE;"`
-	ProductID uint      `json:"productId" gorm:"OnDelete:CASCADE;"`
+	ClientID  int64      `json:"clientId" gorm:"OnDelete:CASCADE;"`
+	ProductID int64      `json:"productId" gorm:"OnDelete:CASCADE;"`
 	Product   Product   `json:"product" gorm:"OnDelete:CASCADE;"`
 	Quantity  int       `json:"quantity"`
 }
